pyproject.toml
[tool.poetry]
name = "ubo-gui"
version = "0.11.0"
description = "GUI sdk for Ubo Pod"
authors = ["Sassan Haradji <sassanh@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "ubo_gui" }]
include = ['ubo_gui/assets/fonts/*']

[tool.poetry.dependencies]
python = "^3.11"
headless-kivy-pi = [
  { version = "^0.7.2", markers = "extra=='default'", extras = [
    'default',
  ] },
  { version = "^0.7.2", markers = "extra=='dev'", extras = [
    'dev',
  ] },
]
python-immutable = "^1.0.2"
qrcode = "^7.4.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.4"
pyright = "^1.1.360"
ruff = "^0.4.2"

[tool.poetry.extras]
dev = ['headless-kivy-pi']
default = ['headless-kivy-pi']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
lint = "ruff check . --unsafe-fixes"
typecheck = "pyright -p pyproject.toml ."
sanity = ["typecheck", "lint"]
download_font.shell = "mkdir -p ubo_gui/assets/fonts; wget https://github.com/google/material-design-icons/raw/master/variablefont/MaterialSymbolsOutlined%5BFILL%2CGRAD%2Copsz%2Cwght%5D.ttf -O ./ubo_gui/assets/fonts/MaterialSymbolsOutlined[FILL,GRAD,opsz,wght].ttf"

[tool.poe.tasks.deploy_to_device]
cmd = 'scripts/deploy.sh'

[tool.ruff]
lint.select = ['ALL']
lint.ignore = ['INP001', 'PLR0911', 'D203', 'D213']
lint.fixable = ['ALL']
lint.unfixable = []

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.format]
quote-style = 'single'

[tool.isort]
profile = "black"

[tool.pyright]
exclude = ['typings']
